Add files via upload
This Project is called "Travis the Security System". It's a program that asks for a users name and runs the name against a list of known users to determine if "Travis" recognizes the users name in the system. If the name is not a recognized name, "Travis" will ask if the user if they'd like to be added to the system. Next time, Travis will recognize the user. Travis will also ask if the user would like to be removed.