Add files via upload
This is a Movie Theater Booking simulator. A Dictionary holds a list of movie, which are the keys. The values are the age requirement and number of seats lefts to see that movie. The program first asks the user which movie they'd like to see. Then asks the age. If user is old enough, it will check how many seats are available and either say "enjoy the movie" or "no more tickets". If User is not old enough, the program will tell the user they aren't old enough to see the movie.